LayoutTests/media/media-source/media-source-seek-during-pending-seek-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(update)
Appended all media segments

Test: Seek during pending seek while paused
EXPECTED (video.currentTime >= 7 == 'true') OK

Test: Seek during pending seek while playing
RUN(video.play())
EVENT(playing)
EXPECTED (video.currentTime >= 8 == 'true') OK
EXPECTED (video.paused == 'false') OK

END OF TEST
